Revert "changed 'long int' definitions to 'int'"
This reverts commit 1263e2727bf958893f4322138b1f53b546ff1a6a.

+++ include/fiwix/msg.h
@@ -40,7 +40,7 @@ struct msqid_ds {
 
 /* message buffer for msgsnd() and msgrcv() */
 struct msgbuf {
-	unsigned int mtype;		/* type of message */
+	long int mtype;			/* type of message */
 	char mtext[1];      		/* message text */
 };
 
@@ -59,7 +59,7 @@ struct msginfo {
 /* one msg structure for each message */
 struct msg {
 	struct msg *msg_next;		/* next message on queue */
-	unsigned int msg_type;
+	long int msg_type;
 	char *msg_spot;			/* message text address */
 	__time_t msg_stime;		/* msgsnd time */
 	short int msg_ts;		/* message text size */
@@ -77,7 +77,7 @@ void msg_release_mq(struct msqid_ds *);
 struct msg *msg_get_new_md(void);
 void msg_release_md(struct msg *);
 int sys_msgsnd(int, const void *, __size_t, int);
-int sys_msgrcv(int, void *, __size_t, unsigned int, int);
+int sys_msgrcv(int, void *, __size_t, long int, int);
 int sys_msgget(key_t, int);
 int sys_msgctl(int, int, struct msqid_ds *);
